Fix weird spacing after experiences link
Browsers appear to be doing something weird to the "about their experiences"
`<a>` tag, where there's a space appended between the tag and the
following period (presumably due to carriage returns around the tag
content.

The space results in odd line wrapping where the period follows on a
separate line when the viewport is narrow, and adds an extra space
between the link text and the period ending the sentence when
the viewport is wide.