Adding Nava's Engineering Apprenticeship program
Hi! We're so grateful for this resource, and wanted to throw our hats in the ring as well. Not sure the most appropriate way to add programs to the list, so opening a simple pull request (feel free to reject and let me know if I should get in touch with someone directly at Thoughtbot). Thank you :)